+++ python/ctranslate2/extensions.py
